+++ CHANGELOG.md
@@ -18,6 +18,10 @@ List all changes in various categories:
 - Handle asynchronous REST APIs (HTTP 202 status code response).
 - Provide ability to send custom GET/POST/PUT/DELETE requests thanks to definition of a pyxelrest service.
 
+### Bug fixes ###
+
+- Auto update settings were not saved.
+
 ## 0.64.1 (2018-02-14) ##
 
 ### Bug fixes ###