+++ CHANGELOG.md
@@ -23,6 +23,10 @@ List all changes in various categories:
 
 - Exceptions were not properly catched on workbook activation event.
 
+### Bug fixes ###
+
+- Auto update settings were not saved.
+
 ## 0.64.1 (2018-02-14) ##
 
 ### Bug fixes ###